o changed the principalType to int o fixed test
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@1589929 13f79535-47bb-0310-9956-ffa450edef68

+++ kerberos-codec/src/main/java/org/apache/directory/server/kerberos/shared/keytab/KeytabDecoder.java
@@ -82,7 +82,7 @@ class KeytabDecoder
     {
         String principalName = getPrincipalName( buffer );
 
-        long principalType = buffer.getInt();
+        int principalType = buffer.getInt();
 
         long time = buffer.getInt();
         KerberosTime timeStamp = new KerberosTime( time * 1000 );
